refactor(#3): reorganize module location
