+++ src/components/templates/default/CertificateSection.tsx
