+++ src/components/templates/default/ContactSection.tsx
@@ -1,10 +1,10 @@
+import { notEmpty } from '@/lib/util'
 import { Link, Text, View } from '@react-pdf/renderer'
 import countries, { type LocaleData } from 'i18n-iso-countries'
 import isoCountries_de from 'i18n-iso-countries/langs/de.json'
 import isoCountries_en from 'i18n-iso-countries/langs/en.json'
 import { useLocale, useTranslations } from 'next-intl'
 import type { PropsWithChildren, ReactNode } from 'react'
-import { notEmpty } from '../lib/util'
 import { Headline } from './Headline'
 import { EarthIcon } from './icons/EarthIcon'
 import { LinkedinIcon } from './icons/LinkedinIcon'
