+++ src/components/templates/default/EducationSection.tsx
