+++ src/components/templates/default/Image.tsx
