+++ src/components/templates/default/icons/EarthIcon.tsx
