+++ src/components/templates/default/icons/MailIcon.tsx
