+++ src/components/templates/default/icons/MapPinIcon.tsx
