+++ src/components/templates/default/styles.ts
