+++ src/components/templates/default/util.ts
@@ -1,3 +1,5 @@
+import { converter, formatHex } from 'culori'
+
 export const locale = 'de-DE'
 
 export function formatDate(
@@ -10,4 +12,10 @@ export function formatDate(
     return
   }
   return Intl.DateTimeFormat(locale, options).format(new Date(date))
+}
+
+const oklchToHex = converter('oklch')
+
+export function colorOclToHex(oclColor: string) {
+  return formatHex(oklchToHex(oclColor))
 }
